backend/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12
django-cors-headers==3.8.0
djongo==1.3.6
psycopg2-binary>=2.8
gunicorn==20.1.0
pymongo==3.12.0
reportlab==3.6.1